+++ src/main/resources/META-INF/resources/WEB-INF/jsp/common/navigation.jspf
@@ -2,7 +2,7 @@
 		<a class="navbar-brand m-1" href="https://courses.in28minutes.com">in28Minutes</a>
 		<div class="collapse navbar-collapse">
 			<ul class="navbar-nav">
-				<li class="nav-item"><a class="nav-link" href="/login">Home</a></li>
+				<li class="nav-item"><a class="nav-link" href="/">Home</a></li>
 				<li class="nav-item"><a class="nav-link" href="/list-todos">Todos</a></li>
 			</ul>
 		</div>
